requirements.txt
python-telegram-bot==22.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
